requirements.txt
python-dotenv==1.0.1
streamlit==1.37.1
langchain==0.2.14
langchain-openai==0.1.22
langchain-ollama==0.1.3
langgraph==0.2.34
tavily-python==0.4.0
